pyproject.toml
[project]
name = "organizationalstructure"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil Korobko",email = "“korobkodaniil396@gmail.com”"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg[binary] (>=3.2.3,<4.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest (>=8.3.4,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
